Solutions/Incredibles/Code/extraction/extraction/requirements.txt
pdfminer.six
PyPDF2